+++ packages/x-charts/src/ChartsLegend/ChartsLegend.tsx
@@ -10,8 +10,9 @@ import { FormattedSeries, SeriesContext } from '../context/SeriesContextProvider
 import { ChartsLegendClasses, getLegendUtilityClass } from './chartsLegendClasses';
 import { DefaultizedProps } from '../models/helpers';
 import { LegendParams } from '../models/seriesType/config';
-import { ChartsText, ChartsTextStyle, getWordsByLines } from '../ChartsText';
+import { ChartsText, ChartsTextStyle } from '../ChartsText';
 import { CardinalDirections } from '../models/layout';
+import { getWordsByLines } from '../internals/getWordsByLines';
 
 export interface ChartsLegendSlots {
   legend?: React.JSXElementConstructor<LegendRendererProps>;
