+++ packages/x-charts/src/ChartsText/ChartsText.tsx
@@ -1,17 +1,6 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { getStringSize } from '../internals/domUtils';
-
-export type ChartsTextBaseline = 'hanging' | 'central' | 'auto';
-
-export interface ChartsTextStyle extends React.CSSProperties {
-  angle?: number;
-  /**
-   * The text baseline
-   * @default 'central'
-   */
-  dominantBaseline?: ChartsTextBaseline;
-}
+import { GetWordsByLinesParams, getWordsByLines } from '../internals/getWordsByLines';
 
 export interface ChartsTextProps
   extends Omit<
@@ -24,29 +13,6 @@ export interface ChartsTextProps
    */
   lineHeight?: number;
   ownerState?: any;
-}
-
-interface GetWordsByLinesParams {
-  /**
-   * Text displayed.
-   */
-  text: string;
-  /**
-   * Style applied to text elements.
-   */
-  style?: ChartsTextStyle;
-  /**
-   * If `true`, the line width is computed.
-   * @default false
-   */
-  needsComputation?: boolean;
-}
-
-export function getWordsByLines({ style, needsComputation, text }: GetWordsByLinesParams) {
-  return text.split('\n').map((subText) => ({
-    text: subText,
-    ...(needsComputation ? getStringSize(subText, style) : { width: 0, height: 0 }),
-  }));
 }
 
 /**
