+++ packages/x-charts/src/ChartsText/index.tsx
@@ -1,2 +1,3 @@
 export { ChartsText } from './ChartsText';
-export type { ChartsTextStyle, ChartsTextProps } from './ChartsText';
+export type { ChartsTextProps } from './ChartsText';
+export type { ChartsTextStyle } from '../internals/getWordsByLines';
