+++ packages/x-charts/src/internals/getWordsByLines.ts
@@ -1,0 +1,35 @@
+import { getStringSize } from './domUtils';
+
+export type ChartsTextBaseline = 'hanging' | 'central' | 'auto';
+
+export interface ChartsTextStyle extends React.CSSProperties {
+  angle?: number;
+  /**
+   * The text baseline
+   * @default 'central'
+   */
+  dominantBaseline?: ChartsTextBaseline;
+}
+
+export interface GetWordsByLinesParams {
+  /**
+   * Text displayed.
+   */
+  text: string;
+  /**
+   * Style applied to text elements.
+   */
+  style?: ChartsTextStyle;
+  /**
+   * If `true`, the line width is computed.
+   * @default false
+   */
+  needsComputation?: boolean;
+}
+
+export function getWordsByLines({ style, needsComputation, text }: GetWordsByLinesParams) {
+  return text.split('\n').map((subText) => ({
+    text: subText,
+    ...(needsComputation ? getStringSize(subText, style) : { width: 0, height: 0 }),
+  }));
+}
